requirements.txt
#-------------------------------------------------------------------------------
# Django
#-------------------------------------------------------------------------------

git+https://github.com/django/django.git@1.7b3

#-------------------------------------------------------------------------------
# Packages
#-------------------------------------------------------------------------------

django-debug-toolbar==1.2
django-haystack==2.1.0
Jinja2==2.7.2
Pillow==2.4.0
psycopg2==2.5.2
pyelasticsearch==0.6.1
Unidecode==0.04.14
watchdog==0.7.1

#-------------------------------------------------------------------------------
# Deployment
#-------------------------------------------------------------------------------

uWSGI==2.0.4
